refactor: rename pass_first/second to index/pseudo
